Rename SignatureController to SignatureService
