+++ lib/steep.rb
@@ -95,6 +95,7 @@ require "steep/server/interaction_worker"
 require "steep/server/master"
 
 require "steep/services/content_change"
+require "steep/services/signature_service"
 
 require "steep/project"
 require "steep/project/pattern"
@@ -108,7 +109,6 @@ require "steep/project/hover_content"
 require "steep/project/completion_provider"
 require "steep/project/stats_calculator"
 
-require "steep/signature_controller"
 require "steep/expectations"
 require "steep/drivers/utils/driver_helper"
 require "steep/drivers/check"
